fix bug

+++ models/resnet2p1d.py
@@ -146,11 +146,11 @@ class ResNet(nn.Module):
                  conv1_t_stride=1,
                  no_max_pool=False,
                  shortcut_type='B',
-                 widen_factor=1,
+                 widen_factor=1.0,
                  n_classes=400):
         super().__init__()
 
-        block_inplanes = [x * widen_factor for x in block_inplanes]
+        block_inplanes = [int(x * widen_factor) for x in block_inplanes]
 
         self.in_planes = block_inplanes[0]
         self.no_max_pool = no_max_pool
